Using command to run jsondiff #11323
